Added CardTest, modified a lot
I modified a lot of classes, getting rid of old method stubs that did
not apply anymore. Added an abstract method performAction to the
abstract Card class, so it would perform all the actions of adding
actions, buys, and coins of the action and treasure cards without having
to have a cast.